code/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.7.2
asttokens==2.4.1
bcrypt==4.1.3
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.16.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==42.0.8
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
Django==4.2.7
executing==2.1.0
Faker==37.0.2
fasttext-wheel==0.9.2
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.0.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jwcrypto==1.5.6
langchain==0.3.20
langchain-core==0.3.41
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langsmith==0.3.11
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.3
openai==1.59.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.0
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
pymssql==2.3.0
pyodbc==5.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.0.6
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.4.4
stack-data==0.6.3
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.3.0
uuid==1.30
wcwidth==0.2.13
Werkzeug==3.1.3
xgboost==3.0.0
zstandard==0.23.0
